# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AI Visualization Portal API",
    "description": "Public REST API for accessing AI model benchmark data",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "email": "support@ai-viz-portal.com"
    }
  },
  "servers": [
    {
      "url": "https://ai-viz-portal.com/api/v1",
      "description": "Production server"
    },
    {
      "url": "https://staging.ai-viz-portal.com/api/v1",
      "description": "Staging server"
    }
  ],
  "paths": {
    "/models": {
      "get": {
        "summary": "List models",
        "description": "Retrieve a list of AI models with optional filters",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "Filter by model name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "benchmark",
            "in": "query",
            "description": "Filter by benchmark",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "minScore",
            "in": "query",
            "description": "Minimum score",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "maxScore",
            "in": "query",
            "description": "Maximum score",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "description": "Sort field",
            "schema": {
              "type": "string",
              "enum": ["name", "score", "date"]
            }
          },
          {
            "name": "sortOrder",
            "in": "query",
            "description": "Sort order",
            "schema": {
              "type": "string",
              "enum": ["asc", "desc"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "score": {
                            "type": "number"
                          },
                          "benchmark": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "type": "integer"
                        },
                        "limit": {
                          "type": "integer"
                        },
                        "total": {
                          "type": "integer"
                        },
                        "totalPages": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/models/{id}": {
      "get": {
        "summary": "Get model details",
        "description": "Retrieve detailed information about a specific model",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Model ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "score": {
                          "type": "number"
                        },
                        "benchmark": {
                          "type": "string"
                        },
                        "description": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Model not found"
          }
        }
      }
    },
    "/benchmarks": {
      "get": {
        "summary": "List benchmarks",
        "description": "Retrieve a list of available benchmarks",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "description": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/rankings": {
      "get": {
        "summary": "Get rankings",
        "description": "Retrieve model rankings for a specific benchmark",
        "parameters": [
          {
            "name": "benchmark",
            "in": "query",
            "description": "Benchmark name",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "rank": {
                            "type": "integer"
                          },
                          "modelId": {
                            "type": "string"
                          },
                          "modelName": {
                            "type": "string"
                          },
                          "score": {
                            "type": "number"
                          },
                          "benchmark": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/compare": {
      "post": {
        "summary": "Compare models",
        "description": "Compare multiple models against each other",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["modelIds"],
                "properties": {
                  "modelIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "benchmark": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "models": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "name": {
                                "type": "string"
                              },
                              "score": {
                                "type": "number"
                              }
                            }
                          }
                        },
                        "benchmark": {
                          "type": "string"
                        },
                        "comparison": {
                          "type": "object",
                          "properties": {
                            "difference": {
                              "type": "number"
                            },
                            "winner": {
                              "type": "string"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}